requirements.txt
streamlit
openai
python-dotenv
qrcode[pil]
pyshorteners
